feat(engine): remove native-mode monkey patch (#350)
* feat(engine): only monkey patch when fallback:

* fix(engine): moving defs around

* fix(engine): fallback descriptors for root

* fix(engine): using root descriptors instead of vm

* fix(engine): lint

* fix(engine): removing unused imports

* fix(engine): not patching events when fallback mode

* fix(engine): configurable true

* fix(engine): lint

* fix(engine): remove only

* fix(engine): configurable true on root descriptors
