+++ packages/lwc-engine/src/framework/dom/event.ts
@@ -1,4 +1,6 @@
 import { getOwnPropertyDescriptor, defineProperties } from "../language";
+import { getCustomElementVM } from "../vm";
+import { addTemplateEventListener, removeTemplateEventListener } from "../events";
 
 // https://dom.spec.whatwg.org/#dom-event-composed
 // This is a very dummy, simple polyfill for composed
@@ -22,3 +24,13 @@ if (!getOwnPropertyDescriptor(Event.prototype, 'composed')) {
 }
 
 export const CustomEvent = (window as any).CustomEvent;
+
+export function addEventListenerPatched(this: EventTarget, type: string, listener: EventListener) {
+    const vm = getCustomElementVM(this as HTMLElement);
+    addTemplateEventListener(vm, type, listener);
+}
+
+export function removeEventListenerPatched(this: EventTarget, type: string, listener: EventListener) {
+    const vm = getCustomElementVM(this as HTMLElement);
+    removeTemplateEventListener(vm, type, listener);
+}
