+++ packages/lwc-engine/src/framework/events.ts
@@ -8,7 +8,7 @@ import {
     parentNodeGetter,
 } from "./dom/node";
 import { VM, OwnerKey, getElementOwnerVM, getCustomElementVM } from "./vm";
-import { isNull, ArraySplice, ArrayIndexOf, create, ArrayPush, isUndefined, isFunction, getOwnPropertyDescriptor, defineProperties } from "./language";
+import { isNull, ArraySplice, ArrayIndexOf, create, ArrayPush, isUndefined, isFunction, getOwnPropertyDescriptor, defineProperties, isTrue } from "./language";
 import { isRendering, vmBeingRendered, invokeEventListener, EventListenerContext, componentEventListenerType } from "./invoker";
 import { patchShadowDomTraversalMethods } from "./dom/traverse";
 
@@ -71,8 +71,10 @@ const eventShadowDescriptors: PropertyDescriptorMap = {
     },
 };
 
-export function patchShadowDomEvent(event: Event) {
-    defineProperties(event, eventShadowDescriptors);
+export function patchShadowDomEvent(vm: VM, event: Event) {
+    if (isTrue(vm.fallback)) {
+        defineProperties(event, eventShadowDescriptors);
+    }
 }
 
 const rootEventListenerMap: WeakMap<EventListener, WrappedListener> = new WeakMap();
@@ -98,7 +100,7 @@ function getWrappedRootListener(vm: VM, listener: EventListener): WrappedListene
                 isChildNode(getRootNode.call(target, event), currentTarget as Node) ||
                 // it is not composed and its is coming from from shadow
                 (composed === false && getRootNode.call(event.target) === currentTarget)) {
-                    patchShadowDomEvent(event);
+                    patchShadowDomEvent(vm, event);
                     invokeEventListener(vm, EventListenerContext.ROOT_LISTENER, listener, undefined, event);
             }
         } as WrappedListener;
@@ -124,7 +126,7 @@ function getWrappedComponentsListener(vm: VM, listener: EventListener): WrappedL
     if (isUndefined(wrappedListener)) {
         wrappedListener = function(event: Event) {
             if (isValidEventForCustomElement(event)) {
-                patchShadowDomEvent(event);
+                patchShadowDomEvent(vm, event);
                 invokeEventListener(vm, EventListenerContext.COMPONENT_LISTENER, listener, undefined, event);
             }
         } as WrappedListener;
