+++ packages/lwc-engine/src/framework/html-element.ts
@@ -1,6 +1,6 @@
 import assert from "./assert";
 import { Component } from "./component";
-import { toString, isObject, freeze, seal, defineProperty, defineProperties, getOwnPropertyNames, ArraySlice, isNull, forEach } from "./language";
+import { toString, isObject, freeze, seal, defineProperty, defineProperties, getOwnPropertyNames, ArraySlice, isNull, forEach, isTrue } from "./language";
 import { addCmpEventListener, removeCmpEventListener } from "./events";
 import {
     getAttribute,
@@ -20,8 +20,28 @@ import { vmBeingConstructed, isBeingConstructed, isRendering, vmBeingRendered } 
 import { getComponentVM, VM } from "./vm";
 import { ArrayReduce, isString, isFunction } from "./language";
 import { observeMutation, notifyMutation } from "./watcher";
-import { CustomEvent } from "./dom/event";
+import { CustomEvent, addEventListenerPatched, removeEventListenerPatched } from "./dom/event";
 import { dispatchEvent } from "./dom/event-target";
+import { lightDomQuerySelector, lightDomQuerySelectorAll } from "./dom/traverse";
+
+const fallbackDescriptors = {
+    querySelector: {
+        value: lightDomQuerySelector,
+        configurable: true,
+    },
+    querySelectorAll: {
+        value: lightDomQuerySelectorAll,
+        configurable: true,
+    },
+    addEventListener: {
+        value: addEventListenerPatched,
+        configurable: true,
+    },
+    removeEventListener: {
+        value: removeEventListenerPatched,
+        configurable: true,
+    },
+};
 
 function getHTMLPropDescriptor(propName: string, descriptor: PropertyDescriptor) {
     const { get, set, enumerable, configurable } = descriptor;
@@ -104,7 +124,7 @@ function LWCElement(this: Component) {
         assert.invariant(vmBeingConstructed.elm instanceof HTMLElement, `Component creation requires a DOM element to be associated to ${vmBeingConstructed}.`);
     }
     const vm = vmBeingConstructed;
-    const { elm, def } = vm;
+    const { elm, def, fallback } = vm;
     const component = this;
     vm.component = component;
     // interaction hooks
@@ -119,6 +139,9 @@ function LWCElement(this: Component) {
     // linking elm and its component with VM
     component[ViewModelReflection] = elm[ViewModelReflection] = vm;
     defineProperties(elm, def.descriptors);
+    if (isTrue(fallback)) {
+        defineProperties(elm, fallbackDescriptors);
+    }
 }
 
 LWCElement.prototype = {
